\name{GN16SData}
\alias{GN16SData}
\docType{data}
\title{
  QIIME-based OTU annotation of the Guerrero Negro hypersaline microbial mat data set.
}
\description{
OTU annotation of the Guerrero Negro hypersaline microbial mat data set. OTUs are predicted using QIIME. QIIME output is imported using the importQIIMEBIOMData function.
}
\usage{data(GN16SData)}
\format{
  The format is a list containing the sample names and OTU predictions of the Guerrero Negro hypersaline microbial mat data set.
}

\source{
Victor Kunin, Jeroen Raes, J Kirk Harris, John R. Spear, Jeffrey J.Walker, Natalia Ivanova, Christian von Mering, Brad M. Bebout, Norman R. Pace, Peer Bork, and Philip Hugenholtz. Millimeter-scale genetic gradients and community-level molecular convergence in a hypersaline microbial mat. Mol Syst Biol, 4:198, 2008.
J Kirk Harris, J Gregory Caporaso, Jeffrey J. Walker, John R. Spear, Nicholas J. Gold, Charles E. Robertson, Philip Hugenholtz, Julia Goodrich, Daniel McDonald, Dan Knights, Paul Marshall, Henry Tufo, Rob Knight, and Norman R. Pace. Phylogenetic stratigraphy in the guerrero negro hypersaline microbial mat. ISME J, 7(1):50–60, Jan 2013.
}
\examples{
data(GN16SData)
}
\keyword{datasets}
