\name{GNMetagenomeData}
\alias{GNMetagenomeData}
\docType{data}
\title{
  Taxonomic abundances for the Guerrero Negro hypersaline microbial mat data set.
}
\description{
Taxonomic abundances for the Guerrero Negro hypersaline microbial mat data set. Taxonomic abundances are computed using Taxy-Pro based on Pfam 27.0 protein domain abundances obtained through the UProC approach. Taxy-Pro data is imported using the importTaxyProData function.
}
\usage{data(GNMetagenomeData)}
\format{
  The format is a list containing the sample names and taxonomic abundances predicted by Taxy-Pro of the Guerrero Negro hypersaline microbial mat data set.
}
\source{
Victor Kunin, Jeroen Raes, J Kirk Harris, John R. Spear, Jeffrey J.Walker, Natalia Ivanova, Christian von Mering, Brad M. Bebout, Norman R. Pace, Peer Bork, and Philip Hugenholtz. Millimeter-scale genetic gradients and community-level molecular convergence in a hypersaline microbial mat. Mol Syst Biol, 4:198, 2008.
J Kirk Harris, J Gregory Caporaso, Jeffrey J. Walker, John R. Spear, Nicholas J. Gold, Charles E. Robertson, Philip Hugenholtz, Julia Goodrich, Daniel McDonald, Dan Knights, Paul Marshall, Henry Tufo, Rob Knight, and Norman R. Pace. Phylogenetic stratigraphy in the guerrero negro hypersaline microbial mat. ISME J, 7(1):50–60, Jan 2013.
}
\examples{
data(GNMetagenomeData)
}
\keyword{datasets}
