% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MoPPro.R
\name{importMoPProReferenceData}
\alias{importMoPProReferenceData}
\title{Import of Tax4Fun and MoP-Pro reference data}
\usage{
importMoPProReferenceData(folder)
}
\arguments{
\item{folder}{(required): a character vector with one character string indicating the folder location of the reference data.}
}
\value{
A list containing the imported reference data for \code{\link{Tax4Fun}} and \code{\link{MoPPro}} prediction.
}
\description{
Import of Tax4Fun and MoP-Pro reference data.
}
\author{
Kathrin P. Asshauer \email{kathrin@gobics.de}
}
\references{
\url{http://gobics.de/kathrin/Tax4Fun/Tax4Fun.html}
}

