% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importQIIMEBiomData.R
\name{importQIIMEBiomData}
\alias{importQIIMEBiomData}
\title{Import of QIIME 16S OTU table(s) in BIOM format}
\usage{
importQIIMEBiomData(inputFiles)
}
\arguments{
\item{inputFiles}{(required): a character vector with one or more character string(s) indicating the file location(s) of the BIOM formatted file(s).}
}
\value{
A list containing the imported and modified QIIME OTU table(s) and sample name(s) which can be used as input for the \code{\link{Tax4Fun}} prediction.
}
\description{
Import of 16S OTU table(s) predicted by QIIME based on the SILVA database and removal of suffixes from taxonomic tags containing uncultured archaeon/organisms/bacterium/crenarchaeote/euryarchaeote.
}
\author{
Kathrin P. Asshauer \email{kathrin@gobics.de}
}
\references{
\url{http://gobics.de/kathrin/Tax4Fun/Tax4Fun.html}
}

