% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importQIIMEData.R
\name{importQIIMEData}
\alias{importQIIMEData}
\title{Import of QIIME 16S OTU table(s) in txt format}
\usage{
importQIIMEData(inputFiles)
}
\arguments{
\item{inputFiles}{(required): a character vector with one or more character string(s) indicating the file location(s) of the txt formatted file(s).}
}
\value{
A list containing the imported and modified QIIME OTU table(s) and sample name(s) which can be used as input for the \code{\link{Tax4Fun}} prediction.
}
\description{
Import of 16S OTU table(s) predicted by QIIME based on the SILVA database and removal of suffixes from taxonomic tags containing uncultured archaeon/organisms/bacterium/crenarchaeote/euryarchaeote.
}
\author{
Kathrin P. Asshauer \email{kathrin@gobics.de}
}
\references{
\url{http://tax4fun.gobics.de/}
}

