% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importQIIMEData.R
\name{importQIIMEDataFile}
\alias{importQIIMEDataFile}
\title{Import of single QIIME 16S OTU table in txt format}
\usage{
importQIIMEDataFile(file)
}
\arguments{
\item{file}{(required): a character vector with one character string indicating the file location of the txt formatted file.}
}
\value{
A list containing the imported and modified QIIME OTU table and sample name which can be used as input for the \code{\link{Tax4Fun}} prediction.
}
\description{
Import of 16S OTU table predicted by QIIME based on the SILVA database and removal of suffixes from taxonomic tags containing uncultured archaeon/organisms/bacterium/crenarchaeote/euryarchaeote.
}
\author{
Kathrin P. Asshauer \email{kathrin@gobics.de}
}
\references{
\url{http://tax4fun.gobics.de/}
}

