% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importSilvaNgsData.R
\name{importSilvaNgsData}
\alias{importSilvaNgsData}
\title{Import of SILVAngs OTU table(s) in csv format}
\usage{
importSilvaNgsData(inputFiles)
}
\arguments{
\item{inputFiles}{(required): a character vector with one or more character string(s) indicating the file location(s) of the csv formatted file(s).}
}
\value{
A list containing the imported SILVAngs OTU table(s) and sample name(s) which can be used as input for the \code{\link{Tax4Fun}} prediction.
}
\description{
Import of 16S OTU table(s) predicted by SILVAngs
}
\author{
Kathrin P. Asshauer \email{kathrin@gobics.de}
}
\references{
\url{http://tax4fun.gobics.de/}
}

