% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importSilvaNgsData.R
\name{importSilvaNgsDataFile}
\alias{importSilvaNgsDataFile}
\title{Import of single SILVAngs file in csv format}
\usage{
importSilvaNgsDataFile(file)
}
\arguments{
\item{file}{(required): a character vector with one character string indicating the file location of the csv formatted file.}
}
\value{
A list containing the imported and modified QIIME OTU table and sample names which can be used as input for the \code{\link{Tax4Fun}} prediction.
}
\description{
Import of single SILVAngs file in csv format.
}
\author{
Kathrin P. Asshauer \email{kathrin@gobics.de}
}
\references{
\url{http://tax4fun.gobics.de/}
}

