% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importTaxyPro.R
\name{importTaxyProData}
\alias{importTaxyProData}
\title{Import of Taxy-Pro profile(s)}
\usage{
importTaxyProData(inputFiles)
}
\arguments{
\item{inputFiles}{(required):  a character vector with one or more character string(s) indicating the file location(s) of the Taxy-Pro csv formatted file(s).}
}
\value{
A list containing the imported Taxy-Pro profile(s) and sample name which can be used as input for the \code{\link{MoPPro}} metabolic prediction  (MoP-Pro approach).
}
\description{
Import of taxonomic profile(s) predicted by Taxy-Pro based on the KEGG bacterial and archaeal organisms.
}
\author{
Kathrin P. Asshauer \email{kathrin@gobics.de}
}
\references{
\url{http://gobics.de/TaxyPro/}
}

