% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importTaxyPro.R
\name{importTaxyProDataFile}
\alias{importTaxyProDataFile}
\title{Import of single Taxy-Pro profile}
\usage{
importTaxyProDataFile(file)
}
\arguments{
\item{file}{(required):  a character vector with one character string indicating the file location of the Taxy-Pro csv formatted file.}
}
\value{
The imported taxonomic profile(s) which can be used as input for the \code{\link{MoPPro}} metabolic prediction (MoP-Pro approach).
}
\description{
Import of single taxonomic profile predicted by Taxy-Pro based on the KEGG bacterial and archaeal organisms.
}
\references{
\url{http://gobics.de/TaxyPro/}
}

